LayoutTests/webanimations/timeline-current-time.html

<script src="../resources/js-test-pre.js"></script>
<script>

description("Check that the AnimationTimeline currentTime property is defined and can be set via internals.");

debug("The document timeline currentTime is null by default.");
shouldBeNull("document.timeline.currentTime");

debug("");
debug("Using internals.setTimelineCurrentTime() can set the document timeline currentTime.");
internals.setTimelineCurrentTime(document.timeline, 1);
shouldBe("document.timeline.currentTime", "1");

debug("");
debug("Using internals.setTimelineCurrentTime() can set the document timeline currentTime to a negative value.");
internals.setTimelineCurrentTime(document.timeline, -1);
shouldBe("document.timeline.currentTime", "-1");

debug("");

</script>
<script src="../resources/js-test-post.js"></script>
