LayoutTests/webanimations/timeline-current-time-expected.txt

Check that the AnimationTimeline currentTime property is defined and can be set via internals.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


The document timeline currentTime is null by default.
PASS document.timeline.currentTime is null

Using internals.setTimelineCurrentTime() can set the document timeline currentTime.
PASS document.timeline.currentTime is 1

Using internals.setTimelineCurrentTime() can set the document timeline currentTime to a negative value.
PASS document.timeline.currentTime is -1

PASS successfullyParsed is true

TEST COMPLETE
